Move more comps to own files
