+++ src/pages/admin/Admin.tsx
@@ -3,7 +3,6 @@ import { Elysia, t } from "elysia";
 import { type Session } from "lucia";
 import { HeaderHtml } from "../../components/header";
 import { LayoutHtml } from "../../components/Layout";
-import { MatchForm } from "../../components/MatchForm";
 import { NavbarHtml } from "../../components/Navbar";
 import { StatsCardHtml } from "../../components/StatsCard";
 import { ctx } from "../../context";
@@ -19,13 +18,13 @@ import {
   getSeasons,
 } from "../../db/queries/seasonQueries";
 import { matches, seasonsTbl } from "../../db/schema";
-import { type Season } from "../../db/schema/season";
 import { isHxRequest, redirect } from "../../lib";
 import { syncIfLocal } from "../../lib/dbHelpers";
-import { EditIcon, TrashIcon } from "../../lib/icons";
-import { type Match } from "../../lib/rating";
-import { cn } from "../../lib/utils";
+import { EditMatchModal } from "./components/EditMatchModal";
+import { EditSeasonModal } from "./components/EditSeasonModal";
+import { ExistingSeasons } from "./components/ExisitngSeasons";
 import { MatchCard } from "./components/MatchCard";
+import { SeasonForm } from "./components/SeasonForm";
 
 export const Admin = new Elysia({
   prefix: "/admin",
@@ -278,254 +277,3 @@ async function page(session: Session | null) {
     </>
   );
 }
-
-interface EditMatchModalProps {
-  match: Match;
-}
-
-const EditMatchModal = ({ match }: EditMatchModalProps) => {
-  return (
-    <div
-      id="edit-match-modal"
-      class={cn(
-        "fixed bottom-0 left-0 right-0 top-0 z-40 backdrop-brightness-50",
-        "flex flex-col items-center justify-center",
-        // TODO: Add animation
-      )}
-      _="on closeEditModal remove me"
-    >
-      <div
-        class="absolute bottom-0 left-0 right-0 top-0 -z-50"
-        _="on click trigger closeEditModal"
-      />
-      <div class="-z-20 w-[80%] max-w-[600px] rounded-md bg-slate-800 p-4 shadow-md lg:p-8">
-        <h1 class="mb-4 text-2xl font-semibold">Edit match</h1>
-        <MatchForm
-          formId={`edit-match-${match.id}-form`}
-          match={match}
-          actionButtons={
-            <div class="mt-3 flex justify-end gap-3">
-              <button
-                type="button"
-                class="rounded-lg bg-red-700 p-2"
-                _="on click trigger closeEditModal"
-              >
-                Cancel
-              </button>
-              <button
-                hx-post="/admin/match"
-                type="submit"
-                class="rounded-lg bg-teal-700 p-2"
-                hx-indicator=".progress-bar"
-                _="on click set my.innerText to 'Saving...' then wait for htmx:afterRequest then set my.innerText to 'Save'"
-              >
-                Save
-              </button>
-            </div>
-          }
-        />
-      </div>
-    </div>
-  );
-};
-
-interface EditSeasonModalProps {
-  season: Season;
-}
-
-const EditSeasonModal = ({ season }: EditSeasonModalProps) => {
-  return (
-    <div
-      id="edit-season-modal"
-      class={cn(
-        "fixed bottom-0 left-0 right-0 top-0 z-40 backdrop-brightness-50",
-        "flex flex-col items-center justify-center",
-        // TODO: Add animation
-      )}
-      _="on closeEditModal remove me"
-    >
-      <div
-        class="absolute bottom-0 left-0 right-0 top-0 -z-50"
-        _="on click trigger closeEditModal"
-      />
-      <div class="-z-20 w-[80%] max-w-[600px] rounded-md bg-slate-800 p-4 text-white shadow-md lg:p-8">
-        <h1 class="mb-4 text-2xl font-semibold">Edit season</h1>
-
-        <SeasonForm
-          formId={`edit-season-${season.id}-form`}
-          season={season}
-          actionButtons={
-            <div class="mt-3 flex justify-end gap-3">
-              <button
-                type="button"
-                class="rounded-lg bg-red-700 p-2"
-                _="on click trigger closeEditModal"
-              >
-                Cancel
-              </button>
-              <button
-                hx-post="/admin/season"
-                type="submit"
-                class="rounded-lg bg-teal-700 p-2"
-                hx-indicator=".progress-bar"
-                _="on click set my.innerText to 'Saving...' then wait for htmx:afterRequest then set my.innerText to 'Save'"
-              >
-                Save
-              </button>
-            </div>
-          }
-        />
-      </div>
-    </div>
-  );
-};
-
-interface SeasonFormProps {
-  formId: string;
-  actionButtons: JSX.Element;
-  amountOfSeasons?: number;
-  season?: Season;
-}
-
-export const SeasonForm = ({
-  formId,
-  actionButtons,
-  amountOfSeasons,
-  season,
-}: SeasonFormProps) => {
-  let startAt = "";
-  let endAt = "";
-
-  // substring 0 - 16 to remove milliseconds and timezone
-  if (season?.startAt) startAt = season.startAt.toISOString().substring(0, 16);
-  if (season?.endAt) endAt = season.endAt.toISOString().substring(0, 16);
-
-  return (
-    <form
-      id={formId}
-      method="post"
-      enctype="multipart/form-data"
-      hx-ext="response-targets"
-      hx-target-400="#errors"
-    >
-      <label for="seasonStart">Start</label>
-      <input
-        id="seasonStart"
-        name="seasonStart"
-        form={formId}
-        class="text-black"
-        type="datetime-local"
-        value={startAt}
-      />
-      <label for="seasonEnd">End</label>
-      <input
-        id="seasonEnd"
-        name="seasonEnd"
-        form={formId}
-        class="text-black"
-        type="datetime-local"
-        value={endAt}
-      />
-      <label for="ratingSystemSelect">Rating system:</label>
-      <select
-        id="ratingSystemSelect"
-        name="ratingSystem"
-        form={formId}
-        class="text-black"
-      >
-        <option
-          value="openskill"
-          selected={season?.ratingSystem === "openskill"}
-        >
-          Openskill
-        </option>
-        <option value="elo" selected={season?.ratingSystem === "elo"}>
-          ELO
-        </option>
-      </select>
-      <label for="seasonName">Season name:</label>
-      <input
-        id="seasonName"
-        name="seasonName"
-        form={formId}
-        class="text-black"
-        type="text"
-        value={
-          season?.name ??
-          `Season ${amountOfSeasons ? amountOfSeasons + 1 : "name"}`
-        }
-      />
-      {actionButtons}
-      <div id="errors" class="text-red-500"></div>
-      {season && <input hidden name="seasonId" value={season.id.toString()} />}
-    </form>
-  );
-};
-
-interface ExistingSeasonsProps {
-  seasons: Season[];
-}
-
-export const ExistingSeasons = ({ seasons }: ExistingSeasonsProps) => {
-  return (
-    <>
-      {seasons.map((season) => (
-        <div>
-          <p>{season.name}</p>
-          <p class="truncate">
-            {season.startAt.toLocaleString("en-US", {
-              hourCycle: "h24",
-              year: "numeric",
-              month: "long",
-              day: "numeric",
-              hour: "2-digit",
-              minute: "2-digit",
-            })}
-          </p>
-          <p class="truncate">
-            {season.endAt.toLocaleString("en-US", {
-              hourCycle: "h24",
-              year: "numeric",
-              month: "long",
-              day: "numeric",
-              hour: "2-digit",
-              minute: "2-digit",
-            })}
-          </p>
-          <p>{season.ratingSystem.toUpperCase()}</p>
-          <div class={cn("mt-auto flex")}>
-            <button
-              hx-get={`admin/season/${season.id}`}
-              hx-target="#mainContainer"
-              hx-swap="afterend"
-              class={cn(
-                "mt-2 flex w-1/2 justify-center gap-3 rounded-l-lg",
-                "bg-teal-700 p-2 hover:bg-teal-700/85",
-              )}
-              _={`on htmx:afterSettle js htmx.process(document.body) end`}
-            >
-              <EditIcon />
-              <p class="hidden sm:block">Edit</p>
-            </button>
-            <button
-              type="button"
-              class={cn(
-                "mt-2 flex w-1/2 justify-center gap-3 rounded-r-lg",
-                "bg-red-700 p-2 hover:bg-red-700/85 disabled:bg-gray-600",
-              )}
-              hx-indicator=".progress-bar"
-              hx-target="#mainContainer"
-              hx-delete={`admin/season/${season.id}`}
-              hx-disabled-elt="this"
-              hx-confirm="Are you sure you wish to delete this season?"
-              _="on htmx:beforeRequest set innerText of <p/> in me to 'Deleting...'"
-            >
-              <TrashIcon />
-              <p class="hidden sm:block">Delete</p>
-            </button>
-          </div>
-        </div>
-      ))}
-    </>
-  );
-};
